# perf(data): optimize struct marshaling/unmarshaling with caching and …